Es wurde versuchsweise ein User-TaskCounter (exit in task.c, load in fat12.c, syscall.c) eingeführt, um die shell steuern zu können. Die Shell geht nur dann in die Eingabeschleife ($>), wenn keine User-Tasks laufen.
git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@100 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/util.c
@@ -137,7 +137,7 @@ char* strcpy(char* dest, const char* src)
    return save;
 }
 
-char* strncpy(char* dest, const char* src, size_t n) // okay?
+char* strncpy(char* dest, const char* src, unsigned int n) // okay?
 {
     if(n != 0)
     {
@@ -157,6 +157,7 @@ char* strncpy(char* dest, const char* src, size_t n) // okay?
      }
      return (dest);
 }
+
 
 /// http://en.wikipedia.org/wiki/Strcat
 char* strcat(char* dest, const char* src)
